Remove cast which was causing an loop in scavenge
The cast to int was causing chunks whose range extends beyond the range
of an int to infinite loop during scavenging, as it repeatedly read the
record closes to int.MaxValue.

+++ src/EventStore.Core/TransactionLog/Chunks/TFChunkScavenger.cs
@@ -399,7 +399,7 @@ namespace EventStore.Core.TransactionLog.Chunks
                     default:
                         throw new ArgumentOutOfRangeException();
                 }
-                result = chunk.TryReadClosestForward((int)result.NextPosition);
+                result = chunk.TryReadClosestForward(result.NextPosition);
             }
         }
